+++ modal/_utils/http_utils.py
@@ -1,18 +1,18 @@
 # Copyright Modal Labs 2022
 import contextlib
-import socket
-import ssl
-from typing import Optional
+from typing import TYPE_CHECKING, Optional
 
-import certifi
-from aiohttp import ClientSession, ClientTimeout, TCPConnector
-from aiohttp.web import Application
-from aiohttp.web_runner import AppRunner, SockSite
+# Note: importing aiohttp seems to take about 100ms, and it's not really necessarily,
+# unless we need to work with blobs. So that's why we import it lazily instead.
+
+if TYPE_CHECKING:
+    from aiohttp import ClientSession
+    from aiohttp.web import Application
 
 from .async_utils import on_shutdown
 
 
-def _http_client_with_tls(timeout: Optional[float]) -> ClientSession:
+def _http_client_with_tls(timeout: Optional[float]) -> "ClientSession":
     """Create a new HTTP client session with standard, bundled TLS certificates.
 
     This is necessary to prevent client issues on some system where Python does
@@ -22,13 +22,18 @@ def _http_client_with_tls(timeout: Optional[float]) -> ClientSession:
     Specifically: the error "unable to get local issuer certificate" when making
     an aiohttp request.
     """
+    import ssl
+
+    import certifi
+    from aiohttp import ClientSession, ClientTimeout, TCPConnector
+
     ssl_context = ssl.create_default_context(cafile=certifi.where())
     connector = TCPConnector(ssl=ssl_context)
     return ClientSession(connector=connector, timeout=ClientTimeout(total=timeout))
 
 
 class ClientSessionRegistry:
-    _client_session: ClientSession
+    _client_session: "ClientSession"
     _client_session_active: bool = False
 
     @staticmethod
@@ -47,9 +52,13 @@ class ClientSessionRegistry:
 
 
 @contextlib.asynccontextmanager
-async def run_temporary_http_server(app: Application):
+async def run_temporary_http_server(app: "Application"):
     # Allocates a random port, runs a server in a context manager
     # This is used in various tests
+    import socket
+
+    from aiohttp.web_runner import AppRunner, SockSite
+
     sock = socket.socket()
     sock.bind(("", 0))
     port = sock.getsockname()[1]
